Moves AsProp utils to Lib
