+++ packages/button/src/Button.tsx
@@ -8,13 +8,8 @@ import {
   useDarkMode,
   useUsingKeyboardContext,
 } from '@leafygreen-ui/leafygreen-provider';
-import {
-  Variant,
-  Size,
-  ButtonProps,
-  isJSXIntrinsicElement,
-  isAsComponent,
-} from './types';
+import { isJSXIntrinsicElement, isAsComponent } from '@leafygreen-ui/lib';
+import { Variant, Size, ButtonProps } from './types';
 import { getClassName, ButtonDataProp } from './styles';
 import { ButtonContent } from './ButtonContent';
